requirements.txt
altgraph==0.17.4
asgiref==3.8.1
attrs==24.2.0
certifi==2024.2.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
distlib==0.3.8
Django==5.0.4
filelock==3.13.4
future==1.0.0
gTTS==2.5.2
h11==0.14.0
huepy==1.2.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
inflect==7.3.1
iniconfig==2.0.0
instabot==0.116.0
instaloader==4.13.1
lxml==5.3.0
mock==5.1.0
more-itertools==10.4.0
moviepy==1.0.3
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
proglog==0.1.10
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pygame==2.6.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyNaCl==1.5.0
pyserial==3.5
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
PySocks==1.7.1
pytest==8.3.3
python-docx==1.1.2
pytz==2024.2
pyusb==1.2.1
pywin32==307
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.25.3
schedule==1.2.2
selenium==4.24.0
setuptools==69.5.1
shiboken6==6.7.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
usb==0.0.83.dev0
virtualenv==20.25.1
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
